# YJIT: Fallback send instructions to vm_sendish