--- a/vm_exec.h
+++ b/vm_exec.h
@@ -169,7 +169,7 @@ default:                        \
 #define THROW_EXCEPTION(exc) return (VALUE)(exc)
 #endif
 
-// Run the interpreter on JIT
+// Run the interpreter from the JIT
 #define VM_EXEC(ec, val) do { \
     if (val == Qundef) { \
         VM_ENV_FLAGS_SET(ec->cfp->ep, VM_FRAME_FLAG_FINISH); \
@@ -177,7 +177,7 @@ default:                        \
     } \
 } while (0)
 
-// Run JIT on the interpreter
+// Run the JIT from the interpreter
 #define JIT_EXEC(ec, val) do { \
     rb_jit_func_t func; \
     if (val == Qundef && (func = jit_compile(ec))) { \