@k0kubun 2023-07-24
I added the Before and After of full YJIT stats as "Details" under "Ratio in YJIT" in the PR description. You can still see it's important to support `block_arg` for optimizing method calls. I think the current counters are good enough for merging this, but I plan to add another counter for counting the number of fallback calls in another PR (I can do this in this PR if you prefer it, but this PR is already complex enough for me).