@maximecb 2023-07-24
>  I plan to add another counter for counting the number of fallback calls in another PR (I can do this in this PR if you prefer it, but this PR is already complex enough for me).

This is a good idea. As we close down exits, it will become more important to track places where we do things that don't exit but could possibly be more optimized. We'll also want to take the habit of profiling more often.